third_party/blink/web_tests/http/tests/devtools/elements/edit/edit-dom-actions-4-expected.txt

Tests that user can mutate DOM by means of elements panel.


Running: testSetUp

Running: testEditInvisibleCharsAsHTML
==== before ====
- <div id="testEditInvisibleCharsAsHTML">
      <div id="node-with-invisible-chars">A&nbsp;B&ensp;C&emsp;D&thinsp;E&zwnj;F&zwj;G&rlm;H&lrm;I</div>
  </div>
<div id="node-with-invisible-chars">A&nbsp;B&ensp;C&emsp;D&thinsp;E&zwnj;F&zwj;G&rlm;H&lrm;I</div>
==== after ====
- <div id="testEditInvisibleCharsAsHTML">
      <div id="node-with-invisible-chars">A#nbsp;B#ensp;C#emsp;D#thinsp;E#zwnj;F#zwj;G#rlm;H#lrm;I</div>
  </div>

Running: testEditScript
==== before ====
- <div id="testEditScript">
    - <script id="node-to-edit-script">
          \n          var i = 0;\n          var j = 5;\n          for (; i < j; ++i) {\n              // Do nothing.\n          }\n\n
      </script>
  </div>
==== after ====
- <div id="testEditScript">
      <script id="node-to-edit-script">var i = 0;\n    var j = 0;\n</script>
  </div>

Running: testEditSVGAttribute
==== before ====
- <div id="testEditSVG">
      <svg id="node-to-edit-svg-attribute" xmlns:xlink="test" width="100">\n        </svg>
  </div>
<svg id="node-to-edit-svg-attribute" xmlns:xlink="test" width="100">
        </svg>
==== after ====
- <div id="testEditSVG">
      <svg id="node-to-edit-svg-attribute" xmlns:xlink="test" width="110">\n        </svg>
  </div>
